Move active directory settings to separate class.
